Moved Action up to top level as per feedback
