+++ builder/actions/git.py
@@ -11,7 +11,7 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from .action import Action
+from action import Action
 
 
 class DownloadSource(Action):
